Real floppy motor control output via User_IO
This pipes the select, motor_on, step and direction outputs from minimig out over the user IO.

I had to disable mt32pi comparability for this to work. There's likely an easy way to get round it but I don't use it at this stage so it's just disabled.

The MiSTer IO outputs 3.3v logic, not 5v which the floppy drives expects. I've had issues with voltage level shifters (one actually killed IO lines on my old DE10) so I am using an Arduino as a level shifter which seems to work fine. I will upload to a new repo soon the code for this but it's super simple.

The floppy drive is powered directly from the USB hub's power output (USB on the back). The arduino is powered from the same place. There are also 10k pullups on the user IO port as I'm using a SNAC breakout board though I suspect they aren't required since the Arduino has internal pullups.

I'll also make a little PCB for this at some point. An Arduino Nano may seem like overkill for this but it's pretty cheap and does a great job.

It also sounds amazing :)